% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{AddressesScopedList}
\alias{AddressesScopedList}
\title{AddressesScopedList Object}
\usage{
AddressesScopedList(AddressesScopedList.warning = NULL,
  AddressesScopedList.warning.data = NULL, addresses = NULL,
  warning = NULL)
}
\arguments{
\item{AddressesScopedList.warning}{The \link{AddressesScopedList.warning} object or list of objects}

\item{AddressesScopedList.warning.data}{The \link{AddressesScopedList.warning.data} object or list of objects}

\item{addresses}{[Output Only] List of addresses contained in this scope}

\item{warning}{[Output Only] Informational warning which replaces the list of addresses when the list is empty}
}
\value{
AddressesScopedList object
}
\description{
AddressesScopedList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other AddressesScopedList functions: \code{\link{AddressesScopedList.warning.data}},
  \code{\link{AddressesScopedList.warning}}
}

