% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{AddressesScopedList.warning}
\alias{AddressesScopedList.warning}
\title{AddressesScopedList.warning Object}
\usage{
AddressesScopedList.warning(AddressesScopedList.warning.data = NULL,
  code = NULL, data = NULL, message = NULL)
}
\arguments{
\item{AddressesScopedList.warning.data}{The \link{AddressesScopedList.warning.data} object or list of objects}

\item{code}{[Output Only] A warning code, if applicable}

\item{data}{[Output Only] Metadata about this warning in key: value format}

\item{message}{[Output Only] A human-readable description of the warning code}
}
\value{
AddressesScopedList.warning object
}
\description{
AddressesScopedList.warning Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
[Output Only] Informational warning which replaces the list of addresses when the list is empty.
}
\seealso{
Other AddressesScopedList functions: \code{\link{AddressesScopedList.warning.data}},
  \code{\link{AddressesScopedList}}
}

