% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{AutoscalersScopedList.warning}
\alias{AutoscalersScopedList.warning}
\title{AutoscalersScopedList.warning Object}
\usage{
AutoscalersScopedList.warning(AutoscalersScopedList.warning.data = NULL,
  code = NULL, data = NULL, message = NULL)
}
\arguments{
\item{AutoscalersScopedList.warning.data}{The \link{AutoscalersScopedList.warning.data} object or list of objects}

\item{code}{[Output Only] A warning code, if applicable}

\item{data}{[Output Only] Metadata about this warning in key: value format}

\item{message}{[Output Only] A human-readable description of the warning code}
}
\value{
AutoscalersScopedList.warning object
}
\description{
AutoscalersScopedList.warning Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
[Output Only] Informational warning which replaces the list of autoscalers when the list is empty.
}
\seealso{
Other AutoscalersScopedList functions: \code{\link{AutoscalersScopedList.warning.data}},
  \code{\link{AutoscalersScopedList}}
}

