% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{AutoscalingPolicyLoadBalancingUtilization}
\alias{AutoscalingPolicyLoadBalancingUtilization}
\title{AutoscalingPolicyLoadBalancingUtilization Object}
\usage{
AutoscalingPolicyLoadBalancingUtilization(utilizationTarget = NULL)
}
\arguments{
\item{utilizationTarget}{Fraction of backend capacity utilization (set in HTTP(s) load balancing configuration) that autoscaler should maintain}
}
\value{
AutoscalingPolicyLoadBalancingUtilization object
}
\description{
AutoscalingPolicyLoadBalancingUtilization Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Configuration parameters of autoscaling based on load balancing.
}

