% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{Backend}
\alias{Backend}
\title{Backend Object}
\usage{
Backend(balancingMode = NULL, capacityScaler = NULL, description = NULL,
  group = NULL, maxConnections = NULL, maxConnectionsPerInstance = NULL,
  maxRate = NULL, maxRatePerInstance = NULL, maxUtilization = NULL)
}
\arguments{
\item{balancingMode}{Specifies the balancing mode for this backend}

\item{capacityScaler}{A multiplier applied to the group's maximum servicing capacity (based on UTILIZATION, RATE or CONNECTION)}

\item{description}{An optional description of this resource}

\item{group}{The fully-qualified URL of a zonal Instance Group resource}

\item{maxConnections}{The max number of simultaneous connections for the group}

\item{maxConnectionsPerInstance}{The max number of simultaneous connections that a single backend instance can handle}

\item{maxRate}{The max requests per second (RPS) of the group}

\item{maxRatePerInstance}{The max requests per second (RPS) that a single backend instance can handle}

\item{maxUtilization}{Used when balancingMode is UTILIZATION}
}
\value{
Backend object
}
\description{
Backend Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Message containing information of one individual backend.
}
