% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functions.R
\name{backendServices.update}
\alias{backendServices.update}
\title{Updates the entire content of the BackendService resource. There are several restrictions and guidelines to keep in mind when updating a backend service. Read  Restrictions and Guidelines for more information.}
\usage{
backendServices.update(BackendService, project, backendService)
}
\arguments{
\item{BackendService}{The \link{BackendService} object to pass to this method}

\item{project}{Project ID for this request}

\item{backendService}{Name of the BackendService resource to update}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/compute)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}

Other BackendService functions: \code{\link{BackendService}},
  \code{\link{backendServices.insert}},
  \code{\link{backendServices.patch}}
}

