% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{BackendServicesScopedList}
\alias{BackendServicesScopedList}
\title{BackendServicesScopedList Object}
\usage{
BackendServicesScopedList(BackendServicesScopedList.warning = NULL,
  BackendServicesScopedList.warning.data = NULL, backendServices = NULL,
  warning = NULL)
}
\arguments{
\item{BackendServicesScopedList.warning}{The \link{BackendServicesScopedList.warning} object or list of objects}

\item{BackendServicesScopedList.warning.data}{The \link{BackendServicesScopedList.warning.data} object or list of objects}

\item{backendServices}{List of BackendServices contained in this scope}

\item{warning}{Informational warning which replaces the list of backend services when the list is empty}
}
\value{
BackendServicesScopedList object
}
\description{
BackendServicesScopedList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other BackendServicesScopedList functions: \code{\link{BackendServicesScopedList.warning.data}},
  \code{\link{BackendServicesScopedList.warning}}
}
