% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functions.R
\docType{package}
\name{compute_googleAuthR}
\alias{compute_googleAuthR}
\alias{compute_googleAuthR-package}
\title{Compute Engine API
Creates and runs virtual machines on Google Cloud Platform.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2016-09-04 00:01:09
filename: /Users/mark/dev/R/autoGoogleAPI/googlecomputev1.auto/R/compute_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
\item https://www.googleapis.com/auth/compute.readonly
\item https://www.googleapis.com/auth/devstorage.full_control
\item https://www.googleapis.com/auth/devstorage.read_only
\item https://www.googleapis.com/auth/devstorage.read_write
}
}

