% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{DeprecationStatus}
\alias{DeprecationStatus}
\title{DeprecationStatus Object}
\usage{
DeprecationStatus(deleted = NULL, deprecated = NULL, obsolete = NULL,
  replacement = NULL, state = NULL)
}
\arguments{
\item{deleted}{An optional RFC3339 timestamp on or after which the deprecation state of this resource will be changed to DELETED}

\item{deprecated}{An optional RFC3339 timestamp on or after which the deprecation state of this resource will be changed to DEPRECATED}

\item{obsolete}{An optional RFC3339 timestamp on or after which the deprecation state of this resource will be changed to OBSOLETE}

\item{replacement}{The URL of the suggested replacement for a deprecated resource}

\item{state}{The deprecation state of this resource}
}
\value{
DeprecationStatus object
}
\description{
DeprecationStatus Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Deprecation status for a public resource.
}
\seealso{
Other DeprecationStatus functions: \code{\link{images.deprecate}}
}

