% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{DiskMoveRequest}
\alias{DiskMoveRequest}
\title{DiskMoveRequest Object}
\usage{
DiskMoveRequest(destinationZone = NULL, targetDisk = NULL)
}
\arguments{
\item{destinationZone}{The URL of the destination zone to move the disk}

\item{targetDisk}{The URL of the target disk to move}
}
\value{
DiskMoveRequest object
}
\description{
DiskMoveRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other DiskMoveRequest functions: \code{\link{projects.moveDisk}}
}

