% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{Firewall.allowed}
\alias{Firewall.allowed}
\title{Firewall.allowed Object}
\usage{
Firewall.allowed()
}
\value{
Firewall.allowed object
}
\description{
Firewall.allowed Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Firewall functions: \code{\link{Firewall}},
  \code{\link{firewalls.insert}},
  \code{\link{firewalls.patch}},
  \code{\link{firewalls.update}}
}
