% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functions.R
\name{firewalls.insert}
\alias{firewalls.insert}
\title{Creates a firewall rule in the specified project using the data included in the request.}
\usage{
firewalls.insert(Firewall, project)
}
\arguments{
\item{Firewall}{The \link{Firewall} object to pass to this method}

\item{project}{Project ID for this request}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/compute)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}

Other Firewall functions: \code{\link{Firewall.allowed}},
  \code{\link{Firewall}}, \code{\link{firewalls.patch}},
  \code{\link{firewalls.update}}
}

