% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{HealthCheckReference}
\alias{HealthCheckReference}
\title{HealthCheckReference Object}
\usage{
HealthCheckReference(healthCheck = NULL)
}
\arguments{
\item{healthCheck}{No description}
}
\value{
HealthCheckReference object
}
\description{
HealthCheckReference Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A full or valid partial URL to a health check. For example, the following are valid URLs:  - https://www.googleapis.com/compute/beta/projects/project-id/global/httpHealthChecks/health-check - projects/project-id/global/httpHealthChecks/health-check - global/httpHealthChecks/health-check
}
