% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{HealthStatus}
\alias{HealthStatus}
\title{HealthStatus Object}
\usage{
HealthStatus(healthState = NULL, instance = NULL, ipAddress = NULL,
  port = NULL)
}
\arguments{
\item{healthState}{Health state of the instance}

\item{instance}{URL of the instance resource}

\item{ipAddress}{The IP address represented by this resource}

\item{port}{The port on the instance}
}
\value{
HealthStatus object
}
\description{
HealthStatus Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

