% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{HttpsHealthCheck}
\alias{HttpsHealthCheck}
\title{HttpsHealthCheck Object}
\usage{
HttpsHealthCheck(checkIntervalSec = NULL, creationTimestamp = NULL,
  description = NULL, healthyThreshold = NULL, host = NULL, id = NULL,
  name = NULL, port = NULL, requestPath = NULL, selfLink = NULL,
  timeoutSec = NULL, unhealthyThreshold = NULL)
}
\arguments{
\item{checkIntervalSec}{How often (in seconds) to send a health check}

\item{creationTimestamp}{[Output Only] Creation timestamp in RFC3339 text format}

\item{description}{An optional description of this resource}

\item{healthyThreshold}{A so-far unhealthy instance will be marked healthy after this many consecutive successes}

\item{host}{The value of the host header in the HTTPS health check request}

\item{id}{[Output Only] The unique identifier for the resource}

\item{name}{Name of the resource}

\item{port}{The TCP port number for the HTTPS health check request}

\item{requestPath}{The request path of the HTTPS health check request}

\item{selfLink}{[Output Only] Server-defined URL for the resource}

\item{timeoutSec}{How long (in seconds) to wait before claiming failure}

\item{unhealthyThreshold}{A so-far healthy instance will be marked unhealthy after this many consecutive failures}
}
\value{
HttpsHealthCheck object
}
\description{
HttpsHealthCheck Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An HttpsHealthCheck resource. This resource defines a template for how individual instances should be checked for health, via HTTPS.
}
\seealso{
Other HttpsHealthCheck functions: \code{\link{httpsHealthChecks.insert}},
  \code{\link{httpsHealthChecks.patch}},
  \code{\link{httpsHealthChecks.update}}
}
