% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{Image.rawDisk}
\alias{Image.rawDisk}
\title{Image.rawDisk Object}
\usage{
Image.rawDisk(containerType = NULL, sha1Checksum = NULL, source = NULL)
}
\arguments{
\item{containerType}{The format used to encode and transmit the block device, which should be TAR}

\item{sha1Checksum}{An optional SHA1 checksum of the disk image before unpackaging; provided by the client when the disk image is created}

\item{source}{The full Google Cloud Storage URL where the disk image is stored}
}
\value{
Image.rawDisk object
}
\description{
Image.rawDisk Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The parameters of the raw disk image.
}
\seealso{
Other Image functions: \code{\link{Image}},
  \code{\link{images.insert}}
}

