% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{Image}
\alias{Image}
\title{Image Object}
\usage{
Image(Image.rawDisk = NULL, archiveSizeBytes = NULL,
  creationTimestamp = NULL, deprecated = NULL, description = NULL,
  diskSizeGb = NULL, family = NULL, guestOsFeatures = NULL, id = NULL,
  imageEncryptionKey = NULL, licenses = NULL, name = NULL,
  rawDisk = NULL, selfLink = NULL, sourceDisk = NULL,
  sourceDiskEncryptionKey = NULL, sourceDiskId = NULL, sourceType = NULL,
  status = NULL)
}
\arguments{
\item{Image.rawDisk}{The \link{Image.rawDisk} object or list of objects}

\item{archiveSizeBytes}{Size of the image tar}

\item{creationTimestamp}{[Output Only] Creation timestamp in RFC3339 text format}

\item{deprecated}{The deprecation status associated with this image}

\item{description}{An optional description of this resource}

\item{diskSizeGb}{Size of the image when restored onto a persistent disk (in GB)}

\item{family}{The name of the image family to which this image belongs}

\item{guestOsFeatures}{A list of features to enable on the guest OS}

\item{id}{[Output Only] The unique identifier for the resource}

\item{imageEncryptionKey}{Encrypts the image using a customer-supplied encryption key}

\item{licenses}{Any applicable license URI}

\item{name}{Name of the resource; provided by the client when the resource is created}

\item{rawDisk}{The parameters of the raw disk image}

\item{selfLink}{[Output Only] Server-defined URL for the resource}

\item{sourceDisk}{URL of the source disk used to create this image}

\item{sourceDiskEncryptionKey}{The customer-supplied encryption key of the source disk}

\item{sourceDiskId}{The ID value of the disk used to create this image}

\item{sourceType}{The type of the image used to create this disk}

\item{status}{[Output Only] The status of the image}
}
\value{
Image object
}
\description{
Image Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An Image resource.
}
\seealso{
Other Image functions: \code{\link{Image.rawDisk}},
  \code{\link{images.insert}}
}
