% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functions.R
\name{images.getFromFamily}
\alias{images.getFromFamily}
\title{Returns the latest image that is part of an image family and is not deprecated.}
\usage{
images.getFromFamily(project, family)
}
\arguments{
\item{project}{Project ID for this request}

\item{family}{Name of the image family to search for}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
\item https://www.googleapis.com/auth/compute.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/compute, https://www.googleapis.com/auth/compute.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}
}

