% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{InstanceAggregatedList.items}
\alias{InstanceAggregatedList.items}
\title{InstanceAggregatedList.items Object}
\usage{
InstanceAggregatedList.items()
}
\value{
InstanceAggregatedList.items object
}
\description{
InstanceAggregatedList.items Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
[Output Only] A map of scoped instance lists.
}
\seealso{
Other InstanceAggregatedList functions: \code{\link{InstanceAggregatedList}}
}
