% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{InstanceGroupList}
\alias{InstanceGroupList}
\title{InstanceGroupList Object}
\usage{
InstanceGroupList(id = NULL, items = NULL, nextPageToken = NULL,
  selfLink = NULL)
}
\arguments{
\item{id}{[Output Only] A unique identifier for this list of instance groups}

\item{items}{A list of instance groups}

\item{nextPageToken}{[Output Only] This token allows you to get the next page of results for list requests}

\item{selfLink}{[Output Only] The URL for this resource type}
}
\value{
InstanceGroupList object
}
\description{
InstanceGroupList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A list of InstanceGroup resources.
}
