% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{InstanceGroupManagerList}
\alias{InstanceGroupManagerList}
\title{InstanceGroupManagerList Object}
\usage{
InstanceGroupManagerList(id = NULL, items = NULL, nextPageToken = NULL,
  selfLink = NULL)
}
\arguments{
\item{id}{[Output Only] A unique identifier for this resource type}

\item{items}{[Output Only] A list of managed instance groups}

\item{nextPageToken}{[Output Only] This token allows you to get the next page of results for list requests}

\item{selfLink}{[Output Only] Server-defined URL for this resource}
}
\value{
InstanceGroupManagerList object
}
\description{
InstanceGroupManagerList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
[Output Only] A list of managed instance groups.
}
