% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functions.R
\name{instanceGroupManagers.abandonInstances}
\alias{instanceGroupManagers.abandonInstances}
\title{Schedules a group action to remove the specified instances from the managed instance group. Abandoning an instance does not delete the instance, but it does remove the instance from any target pools that are applied by the managed instance group. This method reduces the targetSize of the managed instance group by the number of instances that you abandon. This operation is marked as DONE when the action is scheduled even if the instances have not yet been removed from the group. You must separately verify the status of the abandoning action with the listmanagedinstances method.}
\usage{

  instanceGroupManagers.abandonInstances(InstanceGroupManagersAbandonInstancesRequest,
  project, zone, instanceGroupManager)
}
\arguments{
\item{InstanceGroupManagersAbandonInstancesRequest}{The \link{InstanceGroupManagersAbandonInstancesRequest} object to pass to this method}

\item{project}{Project ID for this request}

\item{zone}{The name of the zone where the managed instance group is located}

\item{instanceGroupManager}{The name of the managed instance group}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/compute)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}

Other InstanceGroupManagersAbandonInstancesRequest functions: \code{\link{InstanceGroupManagersAbandonInstancesRequest}}
}
