% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functions.R
\name{instanceGroupManagers.listManagedInstances}
\alias{instanceGroupManagers.listManagedInstances}
\title{Lists all of the instances in the managed instance group. Each instance in the list has a currentAction, which indicates the action that the managed instance group is performing on the instance. For example, if the group is still creating an instance, the currentAction is CREATING. If a previous action failed, the list displays the errors for that failed action.}
\usage{
instanceGroupManagers.listManagedInstances(project, zone, instanceGroupManager,
  filter = NULL, maxResults = NULL, order_by = NULL, pageToken = NULL)
}
\arguments{
\item{project}{Project ID for this request}

\item{zone}{The name of the zone where the managed instance group is located}

\item{instanceGroupManager}{The name of the managed instance group}

\item{filter}{}

\item{maxResults}{}

\item{order_by}{}

\item{pageToken}{}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
\item https://www.googleapis.com/auth/compute.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/compute, https://www.googleapis.com/auth/compute.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}
}
