% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functions.R
\name{instanceGroupManagers.setInstanceTemplate}
\alias{instanceGroupManagers.setInstanceTemplate}
\title{Specifies the instance template to use when creating new instances in this group. The templates for existing instances in the group do not change unless you recreate them.}
\usage{

  instanceGroupManagers.setInstanceTemplate(InstanceGroupManagersSetInstanceTemplateRequest,
  project, zone, instanceGroupManager)
}
\arguments{
\item{InstanceGroupManagersSetInstanceTemplateRequest}{The \link{InstanceGroupManagersSetInstanceTemplateRequest} object to pass to this method}

\item{project}{Project ID for this request}

\item{zone}{The name of the zone where the managed instance group is located}

\item{instanceGroupManager}{The name of the managed instance group}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/compute)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}

Other InstanceGroupManagersSetInstanceTemplateRequest functions: \code{\link{InstanceGroupManagersSetInstanceTemplateRequest}}
}

