% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{InstanceGroupManagersScopedList.warning}
\alias{InstanceGroupManagersScopedList.warning}
\title{InstanceGroupManagersScopedList.warning Object}
\usage{

  InstanceGroupManagersScopedList.warning(InstanceGroupManagersScopedList.warning.data = NULL,
  code = NULL, data = NULL, message = NULL)
}
\arguments{
\item{InstanceGroupManagersScopedList.warning.data}{The \link{InstanceGroupManagersScopedList.warning.data} object or list of objects}

\item{code}{[Output Only] A warning code, if applicable}

\item{data}{[Output Only] Metadata about this warning in key: value format}

\item{message}{[Output Only] A human-readable description of the warning code}
}
\value{
InstanceGroupManagersScopedList.warning object
}
\description{
InstanceGroupManagersScopedList.warning Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
[Output Only] The warning that replaces the list of managed instance groups when the list is empty.
}
\seealso{
Other InstanceGroupManagersScopedList functions: \code{\link{InstanceGroupManagersScopedList.warning.data}},
  \code{\link{InstanceGroupManagersScopedList}}
}
