% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functions.R
\name{instanceGroups.setNamedPorts}
\alias{instanceGroups.setNamedPorts}
\title{Sets the named ports for the specified instance group.}
\usage{
instanceGroups.setNamedPorts(InstanceGroupsSetNamedPortsRequest, project, zone,
  instanceGroup)
}
\arguments{
\item{InstanceGroupsSetNamedPortsRequest}{The \link{InstanceGroupsSetNamedPortsRequest} object to pass to this method}

\item{project}{Project ID for this request}

\item{zone}{The name of the zone where the instance group is located}

\item{instanceGroup}{The name of the instance group where the named ports are updated}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/compute)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}

Other InstanceGroupsSetNamedPortsRequest functions: \code{\link{InstanceGroupsSetNamedPortsRequest}}
}

