% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{InstanceGroupsListInstances}
\alias{InstanceGroupsListInstances}
\title{InstanceGroupsListInstances Object}
\usage{
InstanceGroupsListInstances(id = NULL, items = NULL, nextPageToken = NULL,
  selfLink = NULL)
}
\arguments{
\item{id}{[Output Only] A unique identifier for this list of instances in the specified instance group}

\item{items}{[Output Only] A list of instances and any named ports that are assigned to those instances}

\item{nextPageToken}{[Output Only] This token allows you to get the next page of results for list requests}

\item{selfLink}{[Output Only] The URL for this list of instances in the specified instance groups}
}
\value{
InstanceGroupsListInstances object
}
\description{
InstanceGroupsListInstances Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

