% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{InstanceGroupsScopedList}
\alias{InstanceGroupsScopedList}
\title{InstanceGroupsScopedList Object}
\usage{
InstanceGroupsScopedList(InstanceGroupsScopedList.warning = NULL,
  InstanceGroupsScopedList.warning.data = NULL, instanceGroups = NULL,
  warning = NULL)
}
\arguments{
\item{InstanceGroupsScopedList.warning}{The \link{InstanceGroupsScopedList.warning} object or list of objects}

\item{InstanceGroupsScopedList.warning.data}{The \link{InstanceGroupsScopedList.warning.data} object or list of objects}

\item{instanceGroups}{[Output Only] The list of instance groups that are contained in this scope}

\item{warning}{[Output Only] An informational warning that replaces the list of instance groups when the list is empty}
}
\value{
InstanceGroupsScopedList object
}
\description{
InstanceGroupsScopedList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other InstanceGroupsScopedList functions: \code{\link{InstanceGroupsScopedList.warning.data}},
  \code{\link{InstanceGroupsScopedList.warning}}
}
