% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{InstancesScopedList}
\alias{InstancesScopedList}
\title{InstancesScopedList Object}
\usage{
InstancesScopedList(InstancesScopedList.warning = NULL,
  InstancesScopedList.warning.data = NULL, instances = NULL,
  warning = NULL)
}
\arguments{
\item{InstancesScopedList.warning}{The \link{InstancesScopedList.warning} object or list of objects}

\item{InstancesScopedList.warning.data}{The \link{InstancesScopedList.warning.data} object or list of objects}

\item{instances}{[Output Only] List of instances contained in this scope}

\item{warning}{[Output Only] Informational warning which replaces the list of instances when the list is empty}
}
\value{
InstancesScopedList object
}
\description{
InstancesScopedList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other InstancesScopedList functions: \code{\link{InstancesScopedList.warning.data}},
  \code{\link{InstancesScopedList.warning}}
}
