% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{InstanceWithNamedPorts}
\alias{InstanceWithNamedPorts}
\title{InstanceWithNamedPorts Object}
\usage{
InstanceWithNamedPorts(instance = NULL, namedPorts = NULL, status = NULL)
}
\arguments{
\item{instance}{[Output Only] The URL of the instance}

\item{namedPorts}{[Output Only] The named ports that belong to this instance group}

\item{status}{[Output Only] The status of the instance}
}
\value{
InstanceWithNamedPorts object
}
\description{
InstanceWithNamedPorts Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

