% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{License}
\alias{License}
\title{License Object}
\usage{
License(chargesUseFee = NULL, name = NULL, selfLink = NULL)
}
\arguments{
\item{chargesUseFee}{[Output Only] If true, the customer will be charged license fee for running software that contains this license on an instance}

\item{name}{[Output Only] Name of the resource}

\item{selfLink}{[Output Only] Server-defined URL for the resource}
}
\value{
License object
}
\description{
License Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A license resource.
}

