% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{MachineType}
\alias{MachineType}
\title{MachineType Object}
\usage{
MachineType(MachineType.scratchDisks = NULL, creationTimestamp = NULL,
  deprecated = NULL, description = NULL, guestCpus = NULL, id = NULL,
  imageSpaceGb = NULL, isSharedCpu = NULL, maximumPersistentDisks = NULL,
  maximumPersistentDisksSizeGb = NULL, memoryMb = NULL, name = NULL,
  scratchDisks = NULL, selfLink = NULL, zone = NULL)
}
\arguments{
\item{MachineType.scratchDisks}{The \link{MachineType.scratchDisks} object or list of objects}

\item{creationTimestamp}{[Output Only] Creation timestamp in RFC3339 text format}

\item{deprecated}{[Output Only] The deprecation status associated with this machine type}

\item{description}{[Output Only] An optional textual description of the resource}

\item{guestCpus}{[Output Only] The number of virtual CPUs that are available to the instance}

\item{id}{[Output Only] The unique identifier for the resource}

\item{imageSpaceGb}{[Deprecated] This property is deprecated and will never be populated with any relevant values}

\item{isSharedCpu}{[Output Only] Whether this machine type has a shared CPU}

\item{maximumPersistentDisks}{[Output Only] Maximum persistent disks allowed}

\item{maximumPersistentDisksSizeGb}{[Output Only] Maximum total persistent disks size (GB) allowed}

\item{memoryMb}{[Output Only] The amount of physical memory available to the instance, defined in MB}

\item{name}{[Output Only] Name of the resource}

\item{scratchDisks}{[Output Only] List of extended scratch disks assigned to the instance}

\item{selfLink}{[Output Only] Server-defined URL for the resource}

\item{zone}{[Output Only] The name of the zone where the machine type resides, such as us-central1-a}
}
\value{
MachineType object
}
\description{
MachineType Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Machine Type resource.
}
\seealso{
Other MachineType functions: \code{\link{MachineType.scratchDisks}}
}
