% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{MachineType.scratchDisks}
\alias{MachineType.scratchDisks}
\title{MachineType.scratchDisks Object}
\usage{
MachineType.scratchDisks()
}
\value{
MachineType.scratchDisks object
}
\description{
MachineType.scratchDisks Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other MachineType functions: \code{\link{MachineType}}
}
