% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{MachineTypeList}
\alias{MachineTypeList}
\title{MachineTypeList Object}
\usage{
MachineTypeList(id = NULL, items = NULL, nextPageToken = NULL,
  selfLink = NULL)
}
\arguments{
\item{id}{[Output Only] The unique identifier for the resource}

\item{items}{[Output Only] A list of Machine Type resources}

\item{nextPageToken}{[Output Only] This token allows you to get the next page of results for list requests}

\item{selfLink}{[Output Only] Server-defined URL for this resource}
}
\value{
MachineTypeList object
}
\description{
MachineTypeList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains a list of machine types.
}

