% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{MachineTypesScopedList.warning}
\alias{MachineTypesScopedList.warning}
\title{MachineTypesScopedList.warning Object}
\usage{
MachineTypesScopedList.warning(MachineTypesScopedList.warning.data = NULL,
  code = NULL, data = NULL, message = NULL)
}
\arguments{
\item{MachineTypesScopedList.warning.data}{The \link{MachineTypesScopedList.warning.data} object or list of objects}

\item{code}{[Output Only] A warning code, if applicable}

\item{data}{[Output Only] Metadata about this warning in key: value format}

\item{message}{[Output Only] A human-readable description of the warning code}
}
\value{
MachineTypesScopedList.warning object
}
\description{
MachineTypesScopedList.warning Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
[Output Only] An informational warning that appears when the machine types list is empty.
}
\seealso{
Other MachineTypesScopedList functions: \code{\link{MachineTypesScopedList.warning.data}},
  \code{\link{MachineTypesScopedList}}
}
