% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{ManagedInstanceLastAttempt.errors}
\alias{ManagedInstanceLastAttempt.errors}
\title{ManagedInstanceLastAttempt.errors Object}
\usage{

  ManagedInstanceLastAttempt.errors(ManagedInstanceLastAttempt.errors.errors = NULL,
  errors = NULL)
}
\arguments{
\item{ManagedInstanceLastAttempt.errors.errors}{The \link{ManagedInstanceLastAttempt.errors.errors} object or list of objects}

\item{errors}{[Output Only] The array of errors encountered while processing this operation}
}
\value{
ManagedInstanceLastAttempt.errors object
}
\description{
ManagedInstanceLastAttempt.errors Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
[Output Only] Encountered errors during the last attempt to create or delete the instance.
}
\seealso{
Other ManagedInstanceLastAttempt functions: \code{\link{ManagedInstanceLastAttempt.errors.errors}},
  \code{\link{ManagedInstanceLastAttempt}}
}
