% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{NamedPort}
\alias{NamedPort}
\title{NamedPort Object}
\usage{
NamedPort(name = NULL, port = NULL)
}
\arguments{
\item{name}{The name for this named port}

\item{port}{The port number, which can be a value between 1 and 65535}
}
\value{
NamedPort object
}
\description{
NamedPort Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The named port. For example: .
}

