% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{OperationAggregatedList.items}
\alias{OperationAggregatedList.items}
\title{OperationAggregatedList.items Object}
\usage{
OperationAggregatedList.items()
}
\value{
OperationAggregatedList.items object
}
\description{
OperationAggregatedList.items Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
[Output Only] A map of scoped operation lists.
}
\seealso{
Other OperationAggregatedList functions: \code{\link{OperationAggregatedList}}
}
