% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{PathMatcher}
\alias{PathMatcher}
\title{PathMatcher Object}
\usage{
PathMatcher(defaultService = NULL, description = NULL, name = NULL,
  pathRules = NULL)
}
\arguments{
\item{defaultService}{The full or partial URL to the BackendService resource}

\item{description}{An optional description of this resource}

\item{name}{The name to which this PathMatcher is referred by the HostRule}

\item{pathRules}{The list of path rules}
}
\value{
PathMatcher object
}
\description{
PathMatcher Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A matcher for the path portion of the URL. The BackendService from the longest-matched rule will serve the URL. If no rule was matched, the default service will be used.
}

