% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{PathRule}
\alias{PathRule}
\title{PathRule Object}
\usage{
PathRule(paths = NULL, service = NULL)
}
\arguments{
\item{paths}{The list of path patterns to match}

\item{service}{The URL of the BackendService resource if this rule is matched}
}
\value{
PathRule object
}
\description{
PathRule Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A path-matching rule for a URL. If matched, will use the specified BackendService to handle the traffic arriving at this URL.
}

