% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{Project}
\alias{Project}
\title{Project Object}
\usage{
Project(commonInstanceMetadata = NULL, creationTimestamp = NULL,
  defaultServiceAccount = NULL, description = NULL,
  enabledFeatures = NULL, id = NULL, name = NULL, quotas = NULL,
  selfLink = NULL, usageExportLocation = NULL)
}
\arguments{
\item{commonInstanceMetadata}{Metadata key/value pairs available to all instances contained in this project}

\item{creationTimestamp}{[Output Only] Creation timestamp in RFC3339 text format}

\item{defaultServiceAccount}{[Output Only] Default service account used by VMs running in this project}

\item{description}{An optional textual description of the resource}

\item{enabledFeatures}{Restricted features enabled for use on this project}

\item{id}{[Output Only] The unique identifier for the resource}

\item{name}{The project ID}

\item{quotas}{[Output Only] Quotas assigned to this project}

\item{selfLink}{[Output Only] Server-defined URL for the resource}

\item{usageExportLocation}{The naming prefix for daily usage reports and the Google Cloud Storage bucket where they are stored}
}
\value{
Project object
}
\description{
Project Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Project resource. Projects can only be created in the Google Cloud Platform Console. Unless marked otherwise, values can only be modified in the console.
}

