% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{Quota}
\alias{Quota}
\title{Quota Object}
\usage{
Quota(limit = NULL, metric = NULL, usage = NULL)
}
\arguments{
\item{limit}{[Output Only] Quota limit for this metric}

\item{metric}{[Output Only] Name of the quota metric}

\item{usage}{[Output Only] Current usage of this metric}
}
\value{
Quota object
}
\description{
Quota Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A quotas entry.
}
