% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functions.R
\name{regionBackendServices.getHealth}
\alias{regionBackendServices.getHealth}
\title{Gets the most recent health check results for this regional BackendService.}
\usage{
regionBackendServices.getHealth(ResourceGroupReference, project, region,
  backendService)
}
\arguments{
\item{ResourceGroupReference}{The \link{ResourceGroupReference} object to pass to this method}

\item{project}{}

\item{region}{Name of the region scoping this request}

\item{backendService}{Name of the BackendService resource to which the queried instance belongs}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
\item https://www.googleapis.com/auth/compute.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/compute, https://www.googleapis.com/auth/compute.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}

Other ResourceGroupReference functions: \code{\link{ResourceGroupReference}},
  \code{\link{backendServices.getHealth}}
}
