% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functions.R
\name{regionInstanceGroupManagers.recreateInstances}
\alias{regionInstanceGroupManagers.recreateInstances}
\title{Schedules a group action to recreate the specified instances in the managed instance group. The instances are deleted and recreated using the current instance template for the managed instance group. This operation is marked as DONE when the action is scheduled even if the instances have not yet been recreated. You must separately verify the status of the recreating action with the listmanagedinstances method.}
\usage{

  regionInstanceGroupManagers.recreateInstances(RegionInstanceGroupManagersRecreateRequest,
  project, region, instanceGroupManager)
}
\arguments{
\item{RegionInstanceGroupManagersRecreateRequest}{The \link{RegionInstanceGroupManagersRecreateRequest} object to pass to this method}

\item{project}{Project ID for this request}

\item{region}{Name of the region scoping this request}

\item{instanceGroupManager}{Name of the managed instance group}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/compute)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}

Other RegionInstanceGroupManagersRecreateRequest functions: \code{\link{RegionInstanceGroupManagersRecreateRequest}}
}
