% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{RegionInstanceGroupManagersDeleteInstancesRequest}
\alias{RegionInstanceGroupManagersDeleteInstancesRequest}
\title{RegionInstanceGroupManagersDeleteInstancesRequest Object}
\usage{
RegionInstanceGroupManagersDeleteInstancesRequest(instances = NULL)
}
\arguments{
\item{instances}{The names of one or more instances to delete}
}
\value{
RegionInstanceGroupManagersDeleteInstancesRequest object
}
\description{
RegionInstanceGroupManagersDeleteInstancesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other RegionInstanceGroupManagersDeleteInstancesRequest functions: \code{\link{regionInstanceGroupManagers.deleteInstances}}
}
