% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{RegionInstanceGroupManagersSetTargetPoolsRequest}
\alias{RegionInstanceGroupManagersSetTargetPoolsRequest}
\title{RegionInstanceGroupManagersSetTargetPoolsRequest Object}
\usage{
RegionInstanceGroupManagersSetTargetPoolsRequest(fingerprint = NULL,
  targetPools = NULL)
}
\arguments{
\item{fingerprint}{Fingerprint of the target pools information, which is a hash of the contents}

\item{targetPools}{The URL of all TargetPool resources to which instances in the instanceGroup field are added}
}
\value{
RegionInstanceGroupManagersSetTargetPoolsRequest object
}
\description{
RegionInstanceGroupManagersSetTargetPoolsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other RegionInstanceGroupManagersSetTargetPoolsRequest functions: \code{\link{regionInstanceGroupManagers.setTargetPools}}
}
