% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{RegionInstanceGroupManagersSetTemplateRequest}
\alias{RegionInstanceGroupManagersSetTemplateRequest}
\title{RegionInstanceGroupManagersSetTemplateRequest Object}
\usage{
RegionInstanceGroupManagersSetTemplateRequest(instanceTemplate = NULL)
}
\arguments{
\item{instanceTemplate}{URL of the InstanceTemplate resource from which all new instances will be created}
}
\value{
RegionInstanceGroupManagersSetTemplateRequest object
}
\description{
RegionInstanceGroupManagersSetTemplateRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other RegionInstanceGroupManagersSetTemplateRequest functions: \code{\link{regionInstanceGroupManagers.setInstanceTemplate}}
}
