% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{Route.warnings.data}
\alias{Route.warnings.data}
\title{Route.warnings.data Object}
\usage{
Route.warnings.data()
}
\value{
Route.warnings.data object
}
\description{
Route.warnings.data Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Route functions: \code{\link{Route.warnings}},
  \code{\link{Route}}, \code{\link{routes.insert}}
}

