% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{RouterAggregatedList}
\alias{RouterAggregatedList}
\title{RouterAggregatedList Object}
\usage{
RouterAggregatedList(RouterAggregatedList.items = NULL, id = NULL,
  items = NULL, nextPageToken = NULL, selfLink = NULL)
}
\arguments{
\item{RouterAggregatedList.items}{The \link{RouterAggregatedList.items} object or list of objects}

\item{id}{[Output Only] The unique identifier for the resource}

\item{items}{A map of scoped router lists}

\item{nextPageToken}{[Output Only] This token allows you to get the next page of results for list requests}

\item{selfLink}{[Output Only] Server-defined URL for this resource}
}
\value{
RouterAggregatedList object
}
\description{
RouterAggregatedList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains a list of routers.
}
\seealso{
Other RouterAggregatedList functions: \code{\link{RouterAggregatedList.items}}
}
