% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{RouterBgpPeer}
\alias{RouterBgpPeer}
\title{RouterBgpPeer Object}
\usage{
RouterBgpPeer(advertisedRoutePriority = NULL, interfaceName = NULL,
  ipAddress = NULL, name = NULL, peerAsn = NULL, peerIpAddress = NULL)
}
\arguments{
\item{advertisedRoutePriority}{The priority of routes advertised to this BGP peer}

\item{interfaceName}{Name of the interface the BGP peer is associated with}

\item{ipAddress}{IP address of the interface inside Google Cloud Platform}

\item{name}{Name of this BGP peer}

\item{peerAsn}{Peer BGP Autonomous System Number (ASN)}

\item{peerIpAddress}{IP address of the BGP interface outside Google cloud}
}
\value{
RouterBgpPeer object
}
\description{
RouterBgpPeer Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

