% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{RoutersScopedList}
\alias{RoutersScopedList}
\title{RoutersScopedList Object}
\usage{
RoutersScopedList(RoutersScopedList.warning = NULL,
  RoutersScopedList.warning.data = NULL, routers = NULL, warning = NULL)
}
\arguments{
\item{RoutersScopedList.warning}{The \link{RoutersScopedList.warning} object or list of objects}

\item{RoutersScopedList.warning.data}{The \link{RoutersScopedList.warning.data} object or list of objects}

\item{routers}{List of routers contained in this scope}

\item{warning}{Informational warning which replaces the list of routers when the list is empty}
}
\value{
RoutersScopedList object
}
\description{
RoutersScopedList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other RoutersScopedList functions: \code{\link{RoutersScopedList.warning.data}},
  \code{\link{RoutersScopedList.warning}}
}
