% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{RouterStatusBgpPeerStatus}
\alias{RouterStatusBgpPeerStatus}
\title{RouterStatusBgpPeerStatus Object}
\usage{
RouterStatusBgpPeerStatus(advertisedRoutes = NULL, ipAddress = NULL,
  linkedVpnTunnel = NULL, name = NULL, numLearnedRoutes = NULL,
  peerIpAddress = NULL, state = NULL, status = NULL, uptime = NULL,
  uptimeSeconds = NULL)
}
\arguments{
\item{advertisedRoutes}{Routes that were advertised to the remote BGP peer}

\item{ipAddress}{IP address of the local BGP interface}

\item{linkedVpnTunnel}{URL of the VPN tunnel that this BGP peer controls}

\item{name}{Name of this BGP peer}

\item{numLearnedRoutes}{Number of routes learned from the remote BGP Peer}

\item{peerIpAddress}{IP address of the remote BGP interface}

\item{state}{BGP state as specified in RFC1771}

\item{status}{Status of the BGP peer: {UP, DOWN}}

\item{uptime}{Time this session has been up}

\item{uptimeSeconds}{Time this session has been up, in seconds}
}
\value{
RouterStatusBgpPeerStatus object
}
\description{
RouterStatusBgpPeerStatus Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

