% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{RouterStatusResponse}
\alias{RouterStatusResponse}
\title{RouterStatusResponse Object}
\usage{
RouterStatusResponse(result = NULL)
}
\arguments{
\item{result}{No description}
}
\value{
RouterStatusResponse object
}
\description{
RouterStatusResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
