% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{SerialPortOutput}
\alias{SerialPortOutput}
\title{SerialPortOutput Object}
\usage{
SerialPortOutput(contents = NULL, selfLink = NULL)
}
\arguments{
\item{contents}{[Output Only] The contents of the console output}

\item{selfLink}{[Output Only] Server-defined URL for the resource}
}
\value{
SerialPortOutput object
}
\description{
SerialPortOutput Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An instance's serial console output.
}

