% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{ServiceAccount}
\alias{ServiceAccount}
\title{ServiceAccount Object}
\usage{
ServiceAccount(email = NULL, scopes = NULL)
}
\arguments{
\item{email}{Email address of the service account}

\item{scopes}{The list of scopes to be made available for this service account}
}
\value{
ServiceAccount object
}
\description{
ServiceAccount Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A service account.
}
