% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{SubnetworkAggregatedList.items}
\alias{SubnetworkAggregatedList.items}
\title{SubnetworkAggregatedList.items Object}
\usage{
SubnetworkAggregatedList.items()
}
\value{
SubnetworkAggregatedList.items object
}
\description{
SubnetworkAggregatedList.items Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
[Output] A map of scoped Subnetwork lists.
}
\seealso{
Other SubnetworkAggregatedList functions: \code{\link{SubnetworkAggregatedList}}
}

