% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{SubnetworksExpandIpCidrRangeRequest}
\alias{SubnetworksExpandIpCidrRangeRequest}
\title{SubnetworksExpandIpCidrRangeRequest Object}
\usage{
SubnetworksExpandIpCidrRangeRequest(ipCidrRange = NULL)
}
\arguments{
\item{ipCidrRange}{The IP (in CIDR format or netmask) of internal addresses that are legal on this Subnetwork}
}
\value{
SubnetworksExpandIpCidrRangeRequest object
}
\description{
SubnetworksExpandIpCidrRangeRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other SubnetworksExpandIpCidrRangeRequest functions: \code{\link{subnetworks.expandIpCidrRange}}
}
