% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{SubnetworksScopedList}
\alias{SubnetworksScopedList}
\title{SubnetworksScopedList Object}
\usage{
SubnetworksScopedList(SubnetworksScopedList.warning = NULL,
  SubnetworksScopedList.warning.data = NULL, subnetworks = NULL,
  warning = NULL)
}
\arguments{
\item{SubnetworksScopedList.warning}{The \link{SubnetworksScopedList.warning} object or list of objects}

\item{SubnetworksScopedList.warning.data}{The \link{SubnetworksScopedList.warning.data} object or list of objects}

\item{subnetworks}{List of subnetworks contained in this scope}

\item{warning}{An informational warning that appears when the list of addresses is empty}
}
\value{
SubnetworksScopedList object
}
\description{
SubnetworksScopedList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other SubnetworksScopedList functions: \code{\link{SubnetworksScopedList.warning.data}},
  \code{\link{SubnetworksScopedList.warning}}
}
