% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{TargetHttpProxy}
\alias{TargetHttpProxy}
\title{TargetHttpProxy Object}
\usage{
TargetHttpProxy(creationTimestamp = NULL, description = NULL, id = NULL,
  name = NULL, selfLink = NULL, urlMap = NULL)
}
\arguments{
\item{creationTimestamp}{[Output Only] Creation timestamp in RFC3339 text format}

\item{description}{An optional description of this resource}

\item{id}{[Output Only] The unique identifier for the resource}

\item{name}{Name of the resource}

\item{selfLink}{[Output Only] Server-defined URL for the resource}

\item{urlMap}{URL to the UrlMap resource that defines the mapping from URL to the BackendService}
}
\value{
TargetHttpProxy object
}
\description{
TargetHttpProxy Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A TargetHttpProxy resource. This resource defines an HTTP proxy.
}
\seealso{
Other TargetHttpProxy functions: \code{\link{targetHttpProxies.insert}}
}
