% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{TargetHttpsProxy}
\alias{TargetHttpsProxy}
\title{TargetHttpsProxy Object}
\usage{
TargetHttpsProxy(creationTimestamp = NULL, description = NULL, id = NULL,
  name = NULL, selfLink = NULL, sslCertificates = NULL, urlMap = NULL)
}
\arguments{
\item{creationTimestamp}{[Output Only] Creation timestamp in RFC3339 text format}

\item{description}{An optional description of this resource}

\item{id}{[Output Only] The unique identifier for the resource}

\item{name}{Name of the resource}

\item{selfLink}{[Output Only] Server-defined URL for the resource}

\item{sslCertificates}{URLs to SslCertificate resources that are used to authenticate connections between users and the load balancer}

\item{urlMap}{A fully-qualified or valid partial URL to the UrlMap resource that defines the mapping from URL to the BackendService}
}
\value{
TargetHttpsProxy object
}
\description{
TargetHttpsProxy Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A TargetHttpsProxy resource. This resource defines an HTTPS proxy.
}
\seealso{
Other TargetHttpsProxy functions: \code{\link{targetHttpsProxies.insert}}
}
