% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{TargetInstanceAggregatedList.items}
\alias{TargetInstanceAggregatedList.items}
\title{TargetInstanceAggregatedList.items Object}
\usage{
TargetInstanceAggregatedList.items()
}
\value{
TargetInstanceAggregatedList.items object
}
\description{
TargetInstanceAggregatedList.items Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A map of scoped target instance lists.
}
\seealso{
Other TargetInstanceAggregatedList functions: \code{\link{TargetInstanceAggregatedList}}
}
