% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{TargetInstanceAggregatedList}
\alias{TargetInstanceAggregatedList}
\title{TargetInstanceAggregatedList Object}
\usage{
TargetInstanceAggregatedList(TargetInstanceAggregatedList.items = NULL,
  id = NULL, items = NULL, nextPageToken = NULL, selfLink = NULL)
}
\arguments{
\item{TargetInstanceAggregatedList.items}{The \link{TargetInstanceAggregatedList.items} object or list of objects}

\item{id}{[Output Only] Unique identifier for the resource; defined by the server}

\item{items}{A map of scoped target instance lists}

\item{nextPageToken}{[Output Only] This token allows you to get the next page of results for list requests}

\item{selfLink}{[Output Only] Server-defined URL for this resource}
}
\value{
TargetInstanceAggregatedList object
}
\description{
TargetInstanceAggregatedList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other TargetInstanceAggregatedList functions: \code{\link{TargetInstanceAggregatedList.items}}
}
