% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{TargetInstancesScopedList}
\alias{TargetInstancesScopedList}
\title{TargetInstancesScopedList Object}
\usage{
TargetInstancesScopedList(TargetInstancesScopedList.warning = NULL,
  TargetInstancesScopedList.warning.data = NULL, targetInstances = NULL,
  warning = NULL)
}
\arguments{
\item{TargetInstancesScopedList.warning}{The \link{TargetInstancesScopedList.warning} object or list of objects}

\item{TargetInstancesScopedList.warning.data}{The \link{TargetInstancesScopedList.warning.data} object or list of objects}

\item{targetInstances}{List of target instances contained in this scope}

\item{warning}{Informational warning which replaces the list of addresses when the list is empty}
}
\value{
TargetInstancesScopedList object
}
\description{
TargetInstancesScopedList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other TargetInstancesScopedList functions: \code{\link{TargetInstancesScopedList.warning.data}},
  \code{\link{TargetInstancesScopedList.warning}}
}
