% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{TargetPoolsScopedList}
\alias{TargetPoolsScopedList}
\title{TargetPoolsScopedList Object}
\usage{
TargetPoolsScopedList(TargetPoolsScopedList.warning = NULL,
  TargetPoolsScopedList.warning.data = NULL, targetPools = NULL,
  warning = NULL)
}
\arguments{
\item{TargetPoolsScopedList.warning}{The \link{TargetPoolsScopedList.warning} object or list of objects}

\item{TargetPoolsScopedList.warning.data}{The \link{TargetPoolsScopedList.warning.data} object or list of objects}

\item{targetPools}{List of target pools contained in this scope}

\item{warning}{Informational warning which replaces the list of addresses when the list is empty}
}
\value{
TargetPoolsScopedList object
}
\description{
TargetPoolsScopedList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other TargetPoolsScopedList functions: \code{\link{TargetPoolsScopedList.warning.data}},
  \code{\link{TargetPoolsScopedList.warning}}
}

