% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{TargetSslProxiesSetProxyHeaderRequest}
\alias{TargetSslProxiesSetProxyHeaderRequest}
\title{TargetSslProxiesSetProxyHeaderRequest Object}
\usage{
TargetSslProxiesSetProxyHeaderRequest(proxyHeader = NULL)
}
\arguments{
\item{proxyHeader}{The new type of proxy header to append before sending data to the backend}
}
\value{
TargetSslProxiesSetProxyHeaderRequest object
}
\description{
TargetSslProxiesSetProxyHeaderRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other TargetSslProxiesSetProxyHeaderRequest functions: \code{\link{targetSslProxies.setProxyHeader}}
}
