% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{TestFailure}
\alias{TestFailure}
\title{TestFailure Object}
\usage{
TestFailure(actualService = NULL, expectedService = NULL, host = NULL,
  path = NULL)
}
\arguments{
\item{actualService}{No description}

\item{expectedService}{No description}

\item{host}{No description}

\item{path}{No description}
}
\value{
TestFailure object
}
\description{
TestFailure Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
