% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{UrlMapsValidateRequest}
\alias{UrlMapsValidateRequest}
\title{UrlMapsValidateRequest Object}
\usage{
UrlMapsValidateRequest(resource = NULL)
}
\arguments{
\item{resource}{Content of the UrlMap to be validated}
}
\value{
UrlMapsValidateRequest object
}
\description{
UrlMapsValidateRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other UrlMapsValidateRequest functions: \code{\link{urlMaps.validate}}
}
