% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{VpnTunnelAggregatedList.items}
\alias{VpnTunnelAggregatedList.items}
\title{VpnTunnelAggregatedList.items Object}
\usage{
VpnTunnelAggregatedList.items()
}
\value{
VpnTunnelAggregatedList.items object
}
\description{
VpnTunnelAggregatedList.items Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
[Output Only] A map of scoped vpn tunnel lists.
}
\seealso{
Other VpnTunnelAggregatedList functions: \code{\link{VpnTunnelAggregatedList}}
}
