% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{VpnTunnelAggregatedList}
\alias{VpnTunnelAggregatedList}
\title{VpnTunnelAggregatedList Object}
\usage{
VpnTunnelAggregatedList(VpnTunnelAggregatedList.items = NULL, id = NULL,
  items = NULL, nextPageToken = NULL, selfLink = NULL)
}
\arguments{
\item{VpnTunnelAggregatedList.items}{The \link{VpnTunnelAggregatedList.items} object or list of objects}

\item{id}{[Output Only] The unique identifier for the resource}

\item{items}{[Output Only] A map of scoped vpn tunnel lists}

\item{nextPageToken}{[Output Only] This token allows you to get the next page of results for list requests}

\item{selfLink}{[Output Only] Server-defined URL for this resource}
}
\value{
VpnTunnelAggregatedList object
}
\description{
VpnTunnelAggregatedList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other VpnTunnelAggregatedList functions: \code{\link{VpnTunnelAggregatedList.items}}
}
