% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{VpnTunnelsScopedList}
\alias{VpnTunnelsScopedList}
\title{VpnTunnelsScopedList Object}
\usage{
VpnTunnelsScopedList(VpnTunnelsScopedList.warning = NULL,
  VpnTunnelsScopedList.warning.data = NULL, vpnTunnels = NULL,
  warning = NULL)
}
\arguments{
\item{VpnTunnelsScopedList.warning}{The \link{VpnTunnelsScopedList.warning} object or list of objects}

\item{VpnTunnelsScopedList.warning.data}{The \link{VpnTunnelsScopedList.warning.data} object or list of objects}

\item{vpnTunnels}{List of vpn tunnels contained in this scope}

\item{warning}{Informational warning which replaces the list of addresses when the list is empty}
}
\value{
VpnTunnelsScopedList object
}
\description{
VpnTunnelsScopedList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other VpnTunnelsScopedList functions: \code{\link{VpnTunnelsScopedList.warning.data}},
  \code{\link{VpnTunnelsScopedList.warning}}
}
