% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{SerialPortOutput}
\alias{SerialPortOutput}
\title{SerialPortOutput Object}
\usage{
SerialPortOutput(contents = NULL, next. = NULL, selfLink = NULL,
  start = NULL)
}
\arguments{
\item{contents}{[Output Only] The contents of the console output}

\item{next.}{[Output Only] The position of the next byte of content from the serial console output}

\item{selfLink}{[Output Only] Server-defined URL for this resource}

\item{start}{The starting byte position of the output that was returned}
}
\value{
SerialPortOutput object
}
\description{
SerialPortOutput Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An instance's serial console output.
}
